+++ editor/src/hooks/useInteractiveWorldMinter.ts
@@ -1,14 +1,13 @@
-import { GraphJSON, NodeParameterJSON, NodeParametersJSON, NodeParameterValueJSON } from 'behave-graph';
-import { useCallback, useEffect, useState } from 'react';
+import { GraphJSON, NodeParametersJSON, NodeParameterValueJSON } from 'behave-graph';
+import { useEffect, useState } from 'react';
 import { usePrepareContractWrite, useContractWrite } from 'wagmi';
-import { saveInteractiveWorldToIpfs } from './ipfs/ipfsInteractiveWorldSaver';
 import { abi } from '../contracts/abi';
 import {
   actionNameParamName,
   smartContractInvokedActionName,
   togenGatedAddressParamName,
   tokenGatedParamName,
-} from '../scene/TokenGatedActionInvoker';
+} from '../nodes/smartContracts/TokenGatedActionInvoker';
 
 type TokenizedAction = {
   nodeType: number;
@@ -32,7 +31,9 @@ const getParam = (x: NodeParametersJSON | undefined, paramName: string) => {
 const actionsToSmartContractActions = (behaviorGraph: GraphJSON, contractAddress: string): TokenizedAction[] => {
   const validNodes = behaviorGraph.nodes?.filter((x) => tokenizableActionTypes.includes(x.type));
 
-  const result: TokenizedAction[] = validNodes?.map((x): TokenizedAction => {
+  if (!validNodes) return [];
+
+  const result = validNodes.map((x): TokenizedAction => {
     const parameters = x.parameters;
     const actionName = getParam(parameters, actionNameParamName) as string | undefined;
     const active = !!getParam(parameters, tokenGatedParamName);
@@ -40,7 +41,7 @@ const actionsToSmartContractActions = (behaviorGraph: GraphJSON, contractAddress
 
     if (!actionName) throw new Error(`actionName: ${actionName}  must not be null`);
 
-    return {
+    const inner: TokenizedAction = {
       id: actionName,
       nodeType: 0,
       tokenGateRule: {
@@ -48,6 +49,8 @@ const actionsToSmartContractActions = (behaviorGraph: GraphJSON, contractAddress
         tokenContract: address || (contractAddress as `0x${string}`),
       },
     };
+
+    return inner;
   }) || [contractAddress];
 
   return result;
@@ -57,24 +60,6 @@ const toMintArgs = (cid: string, behaviorGraph: GraphJSON, contractAddress: stri
   cid,
   actionsToSmartContractActions(behaviorGraph, contractAddress),
 ];
-
-export const useSaveSceneToIpfs = ({ modelUrl, behaviorGraph }: { modelUrl: string; behaviorGraph: GraphJSON }) => {
-  const [cid, setCid] = useState<string>();
-  const [saving, setSaving] = useState(false);
-  const saveSceneToIpfs = useCallback(async () => {
-    setSaving(true);
-
-    try {
-      const { cid } = await saveInteractiveWorldToIpfs({ modelUrl, behaviorGraph });
-
-      setCid(cid);
-    } finally {
-      setSaving(false);
-    }
-  }, [modelUrl, behaviorGraph]);
-
-  return { cid, saveSceneToIpfs, saving };
-};
 
 const useInteractiveWorldMinter = ({
   worldCid,
